# api/openapi.yaml
info:
  title: Media API
  version: 0.1.0
tags:
- name: Album
- name: Media
paths:
  /album/:
    post:
      parameters: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AddAlbumOut'
          description: Successful response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
          description: Validation error
      tags:
      - Album
      summary: Add Album
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddAlbumIn'
    delete:
      parameters: []
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
          description: Validation error
      tags:
      - Album
      summary: Delete Album
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeleteAlbumIn'
  /media/:
    get:
      parameters:
      - in: query
        name: media_id
        schema:
          type: string
        required: false
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Media'
          description: Successful response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
          description: Validation error
      tags:
      - Media
      summary: Get Media
    post:
      parameters: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AddMediaOut'
          description: Successful response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
          description: Validation error
      tags:
      - Media
      summary: Add Media
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddMediaIn'
    delete:
      parameters: []
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
          description: Validation error
      tags:
      - Media
      summary: Delete Media
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeleteMediaIn'
  /album/query:
    get:
      parameters:
      - in: query
        name: last_id
        schema:
          type: string
          default: null
          nullable: true
        required: false
      - in: query
        name: limit
        schema:
          type: integer
          default: 30
        required: false
      - in: query
        name: descending
        schema:
          type: boolean
          default: true
        required: false
      - in: query
        name: search
        schema:
          type: string
          default: null
          nullable: true
        required: false
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QueryAlbumsOut'
          description: Successful response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
          description: Validation error
      tags:
      - Album
      summary: Query Albums
  /media/query:
    get:
      parameters:
      - in: query
        name: last_id
        schema:
          type: string
          default: null
          nullable: true
        required: false
      - in: query
        name: limit
        schema:
          type: integer
          default: 30
        required: false
      - in: query
        name: descending
        schema:
          type: boolean
          default: true
        required: false
      - in: query
        name: album_id
        schema:
          type: string
          default: null
          nullable: true
        required: false
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QueryMediaOut'
          description: Successful response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
          description: Validation error
      tags:
      - Media
      summary: Query Media
  /media/add-to-albums:
    post:
      parameters: []
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
          description: Validation error
      tags:
      - Media
      summary: Add Media To Albums
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddMediaToAlbumsIn'
  /media/remove-from-album:
    post:
      parameters: []
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
          description: Validation error
      tags:
      - Media
      summary: Remove Media From Album
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RemoveMediaFromAlbumIn'
openapi: 3.0.3
components:
  schemas:
    ValidationError:
      properties:
        detail:
          type: object
          properties:
            <location>:
              type: object
              properties:
                <field_name>:
                  type: array
                  items:
                    type: string
        message:
          type: string
      type: object
    AddAlbumIn:
      type: object
      properties:
        name:
          type: string
    AddAlbumOut:
      type: object
      properties:
        name:
          type: string
    DeleteAlbumIn:
      type: object
      properties:
        album_ids:
          type: array
          items:
            type: string
    Album:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        thumbnail_path:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    Media:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        uploader:
          type: string
        uploader_id:
          type: string
        uploader_url:
          type: string
        upload_date:
          type: string
        tags:
          type: array
          items:
            type: string
        duration:
          type: integer
        webpage_url:
          type: string
        extractor_key:
          type: string
        media_type:
          type: string
        thumbnail_path:
          type: string
        media_path:
          type: string
        width:
          type: integer
        height:
          type: integer
        albums:
          type: array
          items:
            $ref: '#/components/schemas/Album'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    AddMediaIn:
      type: object
      properties:
        media_url:
          type: string
    AddMediaOut:
      type: object
      properties:
        website:
          type: string
    DeleteMediaIn:
      type: object
      properties:
        media_ids:
          type: array
          items:
            type: string
    QueryAlbumsOut:
      type: object
      properties:
        albums:
          type: array
          items:
            $ref: '#/components/schemas/Album'
        no_more_albums:
          type: boolean
    QueryMediaOut:
      type: object
      properties:
        media:
          type: array
          items:
            $ref: '#/components/schemas/Media'
        no_more_media:
          type: boolean
    AddMediaToAlbumsIn:
      type: object
      properties:
        media_id:
          type: string
        album_ids:
          type: array
          items:
            type: string
    RemoveMediaFromAlbumIn:
      type: object
      properties:
        media_id:
          type: string
        album_id:
          type: string